Move data source schema types
