+++ internal/datasource/common/id/id_schema.go
@@ -1,0 +1,18 @@
+package id
+
+import (
+	datasourceschema "github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+)
+
+func AddToDataSourceSchema(s *datasourceschema.Schema, required bool, description string) {
+	idSchemaAttr := datasourceschema.StringAttribute{}
+	idSchemaAttr.Description = description
+	if required {
+		idSchemaAttr.Required = true
+	} else {
+		idSchemaAttr.Computed = true
+		idSchemaAttr.Required = false
+		idSchemaAttr.Optional = false
+	}
+	s.Attributes["id"] = idSchemaAttr
+}
