+++ internal/resource/common/resourcelink/resource_link_schema.go
@@ -1,7 +1,6 @@
 package resourcelink
 
 import (
-	datasourceschema "github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
@@ -20,20 +19,6 @@ func ToSchema() map[string]schema.Attribute {
 			PlanModifiers: []planmodifier.String{
 				stringplanmodifier.UseStateForUnknown(),
 			},
-		},
-	}
-}
-
-func DataSourceSchema() map[string]datasourceschema.Attribute {
-	return map[string]datasourceschema.Attribute{
-		"id": datasourceschema.StringAttribute{
-			Description: "The ID of the resource.",
-			Required:    true,
-		},
-		"location": datasourceschema.StringAttribute{
-			Description: "A read-only URL that references the resource. If the resource is not currently URL-accessible, this property will be null.",
-			Computed:    true,
-			Optional:    false,
 		},
 	}
 }
